mplemented scheduleshower
