+++ src/main/java/auto_scheduler/ScheduleShower.java
@@ -7,8 +7,7 @@ import java.util.Collection;
 
 public class ScheduleShower implements ScheduleAction {
     ResourcePathProvider pathProvider;
-    ArrayList<CourseInfo> studentSchedule = new ArrayList<>();
-
+    
     public ScheduleShower(ResourcePathProvider pathProvider) {
         this.pathProvider = pathProvider;
     }
@@ -20,27 +19,13 @@ public class ScheduleShower implements ScheduleAction {
         Collection<Integer> s = schedule.get(studentId);
         ArrayList<Integer> sections = new ArrayList<>(s);
         CourseScheduleRepositoryImpl classes = new CourseScheduleRepositoryImpl(path);
-        Collection<CourseInfo> c = classes.getClasses();
-        ArrayList<CourseInfo> courses = new ArrayList<>(c);
+        Collection<CourseInfo> courses = classes.getClasses();
+        ArrayList<CourseInfo> studentSchedule = new ArrayList<>();
         for (Integer sn: sections){
             for (CourseInfo course: courses)
                 if(course.getSectionNumber() == sn)
                     studentSchedule.add(course);
         }
-        System.out.println(this);
-        return null;
-    }
-
-    @Override
-    public String toString() {
-        String classes = "CourseNumber\\CourseName\\SectionNumber\\TimeDay\n";
-        for (CourseInfo course: studentSchedule) {
-            classes += course.getCourseNumber() + "\\" +
-                    course.getCourseName() + "\\" +
-                    course.getSectionNumber() + "\\" +
-                    course.getDayTime() + "\n";
-
-        }
-        return classes;
+        return new ShowScheduleSet(studentId, studentSchedule);
     }
 }
